Fix identifying repository types for fulltext harvesting
(Very) Broken DSpace instances respond with '200 OK' to any request, so
requesting a page that only exists in a given software (e.g. /cgi/handle_404)
will result in a false 200.

Instead we'll try /cgi/users/home and test for 401 (Authorization) in the hope
that other softwares don't occupy that same URL space.